rp/requirements.txt
Django==2.0.8
django-debug-toolbar==1.9.1
mozilla-django-oidc==1.1.2
